+++ raycast-env.d.ts
@@ -1,0 +1,29 @@
+/// <reference types="@raycast/api">
+
+/* 🚧 🚧 🚧
+ * This file is auto-generated from the extension's manifest.
+ * Do not modify manually. Instead, update the `package.json` file.
+ * 🚧 🚧 🚧 */
+
+/* eslint-disable @typescript-eslint/ban-types */
+
+type ExtensionPreferences = {
+  /** PromptHub Server URL - Base URL of PromptHub instance */
+  "serverUrl": string,
+  /** Personal Access Token (PAT) - Optional PAT for PromptHub remote/cloud instances */
+  "apiKey"?: string
+}
+
+/** Preferences accessible in all the extension's commands */
+declare type Preferences = ExtensionPreferences
+
+declare namespace Preferences {
+  /** Preferences accessible in the `index` command */
+  export type Index = ExtensionPreferences & {}
+}
+
+declare namespace Arguments {
+  /** Arguments passed to the `index` command */
+  export type Index = {}
+}
+